Move to seperate classes and prevent double entry to Start
