+++ Terminal.Gui/ViewBase/View.Mouse.cs
@@ -701,113 +701,3 @@ public partial class View // Mouse APIs
 
     private void DisposeMouse () { }
 }
-
-
-/// <summary>
-/// <para>
-/// Handler for raising periodic events while the mouse is held down.
-/// Typically, mouse pointer only needs to be pressed down in a view
-/// to begin this event after which it can be moved elsewhere.
-/// </para>
-/// <para>
-/// Common use cases for this includes holding a button down to increase
-/// a counter (e.g. in <see cref="NumericUpDown"/>).
-/// </para>
-/// </summary>
-public interface IMouseHeldDown : IDisposable
-{
-    // TODO: Guess this should follow the established events type - need to double check what that is.
-    public event EventHandler<CancelEventArgs> MouseIsHeldDownTick;
-
-    void Start ();
-    void Stop ();
-}
-
-class MouseHeldDown : IMouseHeldDown
-{
-    private readonly View _host;
-    private bool _down;
-    private object? _timeout;
-
-    public MouseHeldDown (View host) { _host = host; }
-
-    public event EventHandler<CancelEventArgs>? MouseIsHeldDownTick;
-
-
-    public bool RaiseMouseIsHeldDownTick ()
-    {
-        CancelEventArgs args = new ();
-
-        args.Cancel = OnMouseIsHeldDownTick (args) || args.Cancel;
-
-        if (!args.Cancel && MouseIsHeldDownTick is { })
-        {
-            MouseIsHeldDownTick?.Invoke (this, args);
-        }
-
-        // User event cancelled the mouse held down status so
-        // stop the currently running operation.
-        if (args.Cancel)
-        {
-            Stop ();
-        }
-
-        return args.Cancel;
-    }
-    protected virtual bool OnMouseIsHeldDownTick (CancelEventArgs eventArgs)
-    {
-        return false;
-    }
-
-    public void Start ()
-    {
-        _down = true;
-        Application.GrabMouse (_host);
-
-
-        // Give first tick
-        TickWhileMouseIsHeldDown ();
-
-        // Then periodic ticks
-        _timeout = Application.AddTimeout (TimeSpan.FromMilliseconds (500), TickWhileMouseIsHeldDown);
-
-    }
-
-    private bool TickWhileMouseIsHeldDown ()
-    {
-        if (_down)
-        {
-            RaiseMouseIsHeldDownTick ();
-        }
-        else
-        {
-            Stop ();
-        }
-
-        return _down;
-    }
-
-    public void Stop ()
-    {
-        if (Application.MouseGrabView == _host)
-        {
-            Application.UngrabMouse ();
-        }
-
-        if (_timeout != null)
-        {
-            Application.RemoveTimeout (_timeout);
-        }
-
-        _down = false;
-    }
-
-    public void Dispose ()
-    {
-        if (Application.MouseGrabView == _host)
-        {
-            Stop();
-        }
-    }
-}
-
